blacken, cleanup

+++ lsys/validate.py
@@ -3,13 +3,14 @@ import numpy
 
 
 def axes_object(ax):
-    """ Checks if a value if an Axes. If None, a new one is created.
+    """Checks if a value if an Axes. If None, a new one is created.
     Both the figure and axes are returned (in that order).
 
     """
 
     if ax is None:
         from matplotlib import pyplot
+
         fig, ax = pyplot.subplots()
 
     elif isinstance(ax, matplotlib.axes.Axes):
@@ -22,7 +23,7 @@ def axes_object(ax):
 
 
 def is_np(listlike):
-    if not isinstance(listlike, numpy.ndarray) or listlike.dtype != numpy.float:
-        listlike = numpy.array(listlike, dtype=numpy.float)
+    if not isinstance(listlike, numpy.ndarray) or listlike.dtype != float:
+        listlike = numpy.array(listlike, dtype=float)
 
     return listlike
